# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Baahar API
  version: 0.1.0
  summary: Source-backed public city events
  description: >-
    The contract for Baahar's public discovery feed and protected source
    operations. Public event facts come from verified source versions; missing
    optional values remain null rather than being inferred.
servers:
  - url: http://127.0.0.1:8080
    description: Local API
...
paths:
  /v1/cities:
    get:
      operationId: listCities
      tags: [public]
      summary: List enabled cities
      responses:
        "200":
          description: Enabled city choices
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/City"
        default:
          $ref: "#/components/responses/Problem"
  /v1/events:
    get:
      operationId: listEvents
      tags: [public]
      summary: List current event occurrences for one city and named time window
      description: >-
        Omitting window selects upcoming. Upcoming is bounded to the next 90
        local calendar days. All windows exclude occurrences that ended at or
        before the first page's as_of anchor; ongoing timed and current all-day
        occurrences remain. Pages are ordered by effective start and occurrence
        ID. A signed cursor preserves the first-page anchor and active filters.
      parameters:
        - $ref: "#/components/parameters/CitySlug"
        - name: window
          in: query
          required: false
          schema:
            allOf:
              - $ref: "#/components/schemas/TimeWindow"
            default: upcoming
        - name: category
          in: query
          required: false
          style: form
          explode: false
          schema:
            type: array
            maxItems: 7
            uniqueItems: true
            items:
              $ref: "#/components/schemas/Category"
        - name: free
          in: query
          required: false
          description: When true, include only events explicitly marked free.
          schema:
            type: boolean
        - name: venue
          in: query
          required: false
          description: Exact venue name returned in the feed metadata venue list.
          schema:
            type: string
            minLength: 1
            maxLength: 300
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 1024
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 60
            default: 24
      responses:
        "200":
          description: A stable, cursor-paginated event page
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPage"
        default:
          $ref: "#/components/responses/Problem"
  /v1/ask:
    post:
      operationId: askBaahar
      tags: [public]
      summary: Interpret a short discovery request and return verified matching events
      description: >-
        Converts one stateless request into the same deterministic city, time,
        category, free-entry, and venue filters used by the public feed. The
        interpreter never creates event facts and the response is not stored or
        cached.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskRequest"
      responses:
        "200":
          description: Interpreted filters and a bounded verified preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AskResult"
        default:
          $ref: "#/components/responses/Problem"
  /v1/events/{occurrence_id}:
    parameters:
      - $ref: "#/components/parameters/OccurrenceID"
    get:
      operationId: getEvent
      tags: [public]
      summary: Get the current verified version of an event occurrence
      responses:
        "200":
          description: Current occurrence
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/events/{occurrence_id}/changes:
    parameters:
      - $ref: "#/components/parameters/OccurrenceID"
    get:
      operationId: listEventChanges
      tags: [public]
      summary: List material public changes for an occurrence
      responses:
        "200":
          description: Newest material changes first
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    maxItems: 50
                    items:
                      $ref: "#/components/schemas/EventChange"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
...
  /v1/sources/{source_slug}/summary:
    parameters:
      - name: source_slug
        in: path
        required: true
        schema:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          maxLength: 80
    get:
      operationId: getSourceSummary
      tags: [public]
      summary: Get public freshness and authority information for a source
      responses:
        "200":
          description: Public source summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceSummary"
        "404":
          $ref: "#/components/responses/Problem"
        default:
          $ref: "#/components/responses/Problem"
  /v1/operator/sources:
    get:
      operationId: listOperatorSources
      tags: [operator]
      security:
        - operatorToken: []
      summary: List source health for operators
      responses:
        "200":
          description: Source operations summary
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/OperatorSource"
        default:
          $ref: "#/components/responses/Problem"
  /v1/operator/sources/{source_id}/runs:
    parameters:
      - $ref: "#/components/parameters/SourceID"
    get:
      operationId: listCollectionRuns
      tags: [operator]
      security:
        - operatorToken: []
      summary: List recent collection runs for a source
      responses:
        "200":
          description: Recent runs
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    maxItems: 100
                    items:
                      $ref: "#/components/schemas/CollectionRun"
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: triggerCollectionRun
      tags: [operator]
      security:
        - operatorToken: []
      summary: Enqueue one manual source collection
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Collection was queued or reconciled by idempotency key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionRun"
        default:
          $ref: "#/components/responses/Problem"
  /v1/operator/sources/{source_id}/aliases:
    parameters:
      - $ref: "#/components/parameters/SourceID"
    post:
      operationId: createSourceAlias
      tags: [operator]
      security:
        - operatorToken: []
      summary: Review an incoming ID-less identity as the same occurrence
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSourceAlias"
      responses:
        "201":
          description: Reviewed identity alias created or reconciled idempotently
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceAlias"
        default:
          $ref: "#/components/responses/Problem"
  /v1/operator/runs/{run_id}/replay:
    parameters:
      - $ref: "#/components/parameters/RunID"
    post:
      operationId: replayCollectionRun
      tags: [operator]
      security:
        - operatorToken: []
      summary: Enqueue ingestion replay from an immutable stored artifact
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Replay queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionRun"
        default:
          $ref: "#/components/responses/Problem"
  /v1/operator/incidents/{incident_id}/acknowledge:
    parameters:
      - $ref: "#/components/parameters/IncidentID"
    post:
      operationId: acknowledgeIncident
      tags: [operator]
      security:
        - operatorToken: []
      summary: Acknowledge a source-health incident
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        default:
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    operatorToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
  headers:
    ETag:
      description: Entity tag for conditional requests
      schema:
        type: string
  parameters:
    CitySlug:
      name: city
      in: query
      required: true
      schema:
        $ref: "#/components/schemas/CitySlug"
    OccurrenceID:
      name: occurrence_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SourceID:
      name: source_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RunID:
      name: run_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IncidentID:
      name: incident_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 200
  responses:
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    CitySlug:
      type: string
      minLength: 1
      maxLength: 80
      pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    TimeWindow:
      enum: [upcoming, today, tomorrow, weekend]
    Category:
      enum: [arts, talks, workshops, theatre, music, books, community, other]
    EventStatus:
      enum: [scheduled, cancelled, postponed]
    RegistrationState:
      type: [string, "null"]
      enum: [open, sold_out, closed, not_required, null]
    Freshness:
      enum: [fresh, stale]
    City:
      type: object
      additionalProperties: false
      required: [slug, name, timezone, accent]
      properties:
        slug:
          $ref: "#/components/schemas/CitySlug"
        name:
          type: string
          minLength: 1
          maxLength: 100
        timezone:
          type: string
          minLength: 1
          maxLength: 100
        accent:
          type: string
          minLength: 1
          maxLength: 40
    Timing:
      type: object
      additionalProperties: false
      required: [start_date, starts_at, end_date, ends_at, precision, timezone]
      properties:
        start_date:
          type: string
          format: date
        starts_at:
          type: [string, "null"]
          format: date-time
        end_date:
          type: [string, "null"]
          format: date
        ends_at:
          type: [string, "null"]
          format: date-time
        precision:
          enum: [timed, date]
        timezone:
          const: Asia/Kolkata
      allOf:
        - if:
            properties:
              precision:
                const: timed
            required: [precision]
          then:
            properties:
              starts_at:
                type: string
                format: date-time
          else:
            properties:
              starts_at:
                type: "null"
              ends_at:
                type: "null"
        - if:
            properties:
              ends_at:
                type: string
            required: [ends_at]
          then:
            properties:
              end_date:
                type: string
                format: date
    Venue:
      type: [object, "null"]
      additionalProperties: false
      required: [name, address]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 300
        address:
          type: [string, "null"]
          maxLength: 1000
    Pricing:
      type: object
      additionalProperties: false
      required: [is_free, minimum_minor, maximum_minor, currency]
      properties:
        is_free:
          type: [boolean, "null"]
        minimum_minor:
          type: [integer, "null"]
          minimum: 0
        maximum_minor:
          type: [integer, "null"]
          minimum: 0
        currency:
          type: [string, "null"]
          enum: [INR, null]
      allOf:
        - if:
            anyOf:
              - properties:
                  minimum_minor:
                    type: integer
                required: [minimum_minor]
              - properties:
                  maximum_minor:
                    type: integer
                required: [maximum_minor]
          then:
            properties:
              currency:
                const: INR
              is_free:
                const: false
        - if:
            properties:
              is_free:
                const: true
            required: [is_free]
          then:
            properties:
              minimum_minor:
                type: "null"
              maximum_minor:
                type: "null"
              currency:
                type: "null"
        - if:
            properties:
              maximum_minor:
                type: integer
            required: [maximum_minor]
          then:
            properties:
              minimum_minor:
                type: integer
        - if:
            properties:
              minimum_minor:
                type: "null"
              maximum_minor:
                type: "null"
            required: [minimum_minor, maximum_minor]
          then:
            properties:
              currency:
                type: "null"
    Registration:
      type: object
      additionalProperties: false
      required: [url, state]
      properties:
        url:
          type: [string, "null"]
          format: uri
        state:
          $ref: "#/components/schemas/RegistrationState"
    SourceReference:
      type: object
      additionalProperties: false
      required: [slug, name, url, host, freshness]
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        name:
          type: string
        url:
          type: string
          format: uri
        host:
          type: string
          format: hostname
        freshness:
          $ref: "#/components/schemas/Freshness"
    EventSummary:
      type: object
      additionalProperties: false
      required:
        - id
        - slug
        - city
        - title
        - category
        - timing
        - venue
        - pricing
        - registration
        - status
        - image_url
        - source
        - last_checked_at
        - change_kind
        - language
        - age_note
        - accessibility_note
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          maxLength: 120
        city:
          $ref: "#/components/schemas/City"
        title:
          type: string
          minLength: 1
          maxLength: 300
        category:
          $ref: "#/components/schemas/Category"
        timing:
          $ref: "#/components/schemas/Timing"
        venue:
          $ref: "#/components/schemas/Venue"
        pricing:
          $ref: "#/components/schemas/Pricing"
        registration:
          $ref: "#/components/schemas/Registration"
        status:
          $ref: "#/components/schemas/EventStatus"
        image_url:
          type: [string, "null"]
          format: uri
        source:
          $ref: "#/components/schemas/SourceReference"
        last_checked_at:
          type: string
          format: date-time
        change_kind:
          type: [string, "null"]
          enum: [new, updated, null]
        language:
          type: array
          maxItems: 12
          uniqueItems: true
          items:
            type: string
        age_note:
          type: [string, "null"]
          maxLength: 300
        accessibility_note:
          type: [string, "null"]
          maxLength: 1000
    EventDetail:
      $ref: "#/components/schemas/EventSummary"
    EventPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor, meta]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/EventSummary"
        next_cursor:
          type: [string, "null"]
          description: Signed continuation cursor; non-null exactly when meta.has_more is true.
        meta:
          type: object
          additionalProperties: false
          required:
            [
              city,
              window,
              result_count,
              source_count,
              last_checked_at,
              venues,
              page_size,
              has_more,
              as_of,
            ]
          properties:
            city:
              $ref: "#/components/schemas/City"
            window:
              $ref: "#/components/schemas/TimeWindow"
            result_count:
              type: integer
              minimum: 0
              description: Total occurrences matching the anchored filters, not only this page.
            source_count:
              type: integer
              minimum: 0
            last_checked_at:
              type: [string, "null"]
              format: date-time
            venues:
              type: array
              maxItems: 300
              uniqueItems: true
              description: Exact current venue names available for this city and filter window.
              items:
                type: string
                minLength: 1
                maxLength: 300
            page_size:
              type: integer
              minimum: 0
              maximum: 60
              description: Number of items returned in this page.
            has_more:
              type: boolean
              description: Whether a continuation page is available through next_cursor.
            as_of:
              type: string
              format: date-time
              description: Stable request-time anchor carried by every cursor in this result set.
    AskRequest:
      type: object
      additionalProperties: false
      required: [city, query]
      properties:
        city:
          $ref: "#/components/schemas/CitySlug"
        query:
          type: string
          minLength: 1
          maxLength: 280
    AskInterpretation:
      type: object
      additionalProperties: false
      required: [window, categories, explicitly_free, venue, assisted]
      properties:
        window:
          $ref: "#/components/schemas/TimeWindow"
        categories:
          type: array
          maxItems: 3
          uniqueItems: true
          items:
            $ref: "#/components/schemas/Category"
        explicitly_free:
          type: boolean
        venue:
          type: [string, "null"]
          maxLength: 300
        assisted:
          type: boolean
          description: True when the configured language model produced the validated interpretation.
    AskResult:
      type: object
      additionalProperties: false
      required: [interpretation, items, result_count, as_of]
      properties:
        interpretation:
          $ref: "#/components/schemas/AskInterpretation"
        items:
          type: array
          maxItems: 6
          items:
            $ref: "#/components/schemas/EventSummary"
        result_count:
          type: integer
          minimum: 0
        as_of:
          type: string
          format: date-time
    EventChange:
      type: object
      additionalProperties: false
      required: [id, kind, changed_fields, changed_at]
      properties:
        id:
          type: string
          format: uuid
        kind:
          enum: [updated, cancelled, postponed]
        changed_fields:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            enum:
              - timing
              - venue
              - pricing
              - registration
              - status
        changed_at:
          type: string
          format: date-time
    SourceSummary:
      type: object
      additionalProperties: false
      required: [slug, name, official_url, city, freshness, last_healthy_at]
      properties:
        slug:
          type: string
        name:
          type: string
        official_url:
          type: string
          format: uri
        city:
          $ref: "#/components/schemas/City"
        freshness:
          $ref: "#/components/schemas/Freshness"
        last_healthy_at:
          type: [string, "null"]
          format: date-time
    OperatorSource:
      type: object
      additionalProperties: false
      required:
        - id
        - slug
        - name
        - official_url
        - city
        - freshness
        - last_healthy_at
        - collector_id
        - schema_version
        - publication_state
        - next_due_at
        - latest_run
        - active_incident
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        official_url:
          type: string
          format: uri
        city:
          $ref: "#/components/schemas/City"
        freshness:
          $ref: "#/components/schemas/Freshness"
        last_healthy_at:
          type: [string, "null"]
          format: date-time
        collector_id:
          type: string
          pattern: "^c_[A-Za-z0-9_-]+$"
        schema_version:
          type: string
          minLength: 1
          maxLength: 100
        publication_state:
          enum: [active, frozen, disabled]
        next_due_at:
          type: [string, "null"]
          format: date-time
        latest_run:
          oneOf:
            - $ref: "#/components/schemas/CollectionRun"
            - type: "null"
        active_incident:
          oneOf:
            - $ref: "#/components/schemas/Incident"
            - type: "null"
    CollectionRun:
      type: object
      additionalProperties: false
      required:
        - id
        - source_id
        - external_collection_id
        - status
        - triggered_at
        - completed_at
        - accepted_count
        - quarantined_count
        - health_code
      properties:
        id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        external_collection_id:
          type: [string, "null"]
        status:
          enum:
            [
              queued,
              triggering,
              collecting,
              validating,
              published,
              rejected,
              failed,
            ]
        triggered_at:
          type: string
          format: date-time
        completed_at:
          type: [string, "null"]
          format: date-time
        accepted_count:
          type: integer
          minimum: 0
        quarantined_count:
          type: integer
          minimum: 0
        health_code:
          type: [string, "null"]
          maxLength: 100
    Incident:
      type: object
      additionalProperties: false
      required:
        [id, source_id, run_id, code, state, created_at, acknowledged_at]
      properties:
        id:
          type: string
          format: uuid
        source_id:
          type: string
          format: uuid
        run_id:
          type: [string, "null"]
          format: uuid
        code:
          type: string
          maxLength: 100
        state:
          enum: [open, acknowledged, resolved]
        created_at:
          type: string
          format: date-time
        acknowledged_at:
          type: [string, "null"]
          format: date-time
    CreateSourceAlias:
      type: object
      additionalProperties: false
      required: [incoming_identity, occurrence_id, reason]
      properties:
        incoming_identity:
          type: string
          pattern: "^[0-9a-f]{64}$"
        occurrence_id:
          type: string
          format: uuid
        reason:
          type: string
          minLength: 1
          maxLength: 1000
    SourceAlias:
      type: object
      additionalProperties: false
      required:
        [
          source_id,
          incoming_identity,
          occurrence_id,
          merged_occurrence_id,
          reason,
          created_at,
        ]
      properties:
        source_id:
          type: string
          format: uuid
        incoming_identity:
          type: string
          pattern: "^[0-9a-f]{64}$"
        occurrence_id:
          type: string
          format: uuid
        merged_occurrence_id:
          type: [string, "null"]
          format: uuid
        reason:
          type: string
          minLength: 1
          maxLength: 1000
        created_at:
          type: string
          format: date-time
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: "^[a-z0-9_]+$"